feat: centralize all prompts inside the prompts.ts file
