+++ app/api/summary+api.ts
@@ -1,15 +1,7 @@
 import { google } from '@ai-sdk/google';
 import { TextUIPart } from '@ai-sdk/ui-utils';
 import { CoreMessage, generateText, GenerateTextResult, ToolSet } from 'ai';
-import dedent from 'dedent';
-
-const SYSTEM_PROMPT = dedent`
-    You are a personalized health assistant summarizing health-related conversations.
-    Your ONE AND ONLY job is to generate short conversations summaries of the discussion between an AI agent and a user.
-    The summary must be concise and informative, focusing on health topics, concerns, and advice discussed.
-    Its length must be between 10 and 30 words. The less, the better.
-    Avoid any markdown, only plain text.
-`;
+import { getSummaryPrompt } from '@/utils/prompts';
 
 
 export async function POST(req: Request) {
@@ -25,7 +17,7 @@ export async function POST(req: Request) {
         result = await generateText({
             model: google('gemini-2.0-flash-001'),
             messages,
-            system: SYSTEM_PROMPT,
+            system: getSummaryPrompt(),
             maxTokens: 200,
         });
     } catch (e) {
