+++ app/api/title+api.ts
@@ -1,16 +1,7 @@
 import { google } from '@ai-sdk/google';
 import { TextUIPart } from '@ai-sdk/ui-utils';
 import { CoreMessage, generateText, GenerateTextResult, ToolSet } from 'ai';
-
-const SYSTEM_PROMPT = [
-    'You are a personalized health assistant.',
-    'Your ONE AND ONLY job is to generate short conversation titles based on a user prompt and an answer.',
-    'The title must be concise and informative.',
-    'Its length must be between 3 and 7 words. The less, the better.',
-    'Try to be subjective to the situation.',
-    'Always generate the title in the language of the first prompt and answer. Default to English.',
-    'Avoid any markdown, only plain text.',
-].join(' ');
+import { getTitlePrompt } from '@/utils/prompts';
 
 
 export async function POST(req: Request) {
@@ -26,7 +17,7 @@ export async function POST(req: Request) {
         result = await generateText({
             model: google('gemini-2.0-flash-001'),
             messages,
-            system: SYSTEM_PROMPT,
+            system: getTitlePrompt(),
             maxTokens: 12,
         });
     } catch (e) {
