+++ utils/ai.ts
@@ -1,7 +1,6 @@
 import { ToolSet, UIMessage } from 'ai';
 import { fetch as expoFetch } from 'expo/fetch';
 import { z } from 'zod';
-import dedent from 'dedent';
 import { generateAPIUrl } from '@/utils/endpoints';
 
 
@@ -93,37 +92,3 @@ export const tools = {
         }),
     },
 } satisfies ToolSet;
-
-
-const getCurrentDateFormatted = () => new Date().toLocaleString('en', {
-    day: 'numeric',
-    month: 'long',
-    weekday: 'long',
-    year: 'numeric',
-});
-
-export interface SystemPromptOptions {
-    tone: string;
-    adviceMode: string;
-}
-
-export const getSystemPrompt = (options: SystemPromptOptions) => dedent`
-    You are a ${options.tone} personalized health assistant.
-    Your role is to help the user with his health and lifestyle.
-    For that, you have access to his health data, like steps count, sleep time and exercise.
-    You must give the user advice ${options.adviceMode}
-    Advice can be to sleep more, exercise more, or to take care of his health in general.
-    Try to always relate the advice to the data you have, like a doctor or a health coach would do.
-    Don't answer with markdown, only plain text. Don't even use markings like **. For lists, use dashes.
-    Always answer in the same language as the question, no matter what. Default to English.
-    Always format properly durations, like 1 hour 30 minutes instead of 90 minutes.
-    You can chain tools together. For instance, get steps count and then display it to the user.
-    Try to display information in a graph when relevant.
-    Always display periods between 4 and 14 days on the graphs, include the subset of relevant periods in that range.
-    If you have already called a tool once, don't call it again with the same parameters, use the result from the first call.
-    Always respond some text, never tools invocations alone. Interpret and explain the data.
-    Don't show the graph and say "see by yourself", give a text answer to the question.
-    Don't enumerate data to the user (like saying day by day numbers), prefer to show graphs, summarize and interpret the data.
-    You only have access to the last 30 days of data.
-    For your information, today is ${getCurrentDateFormatted()}.
-`;
